refactor: rename module test_spec
rename

- test_spec -> testcase
- test-spec -> testcase
