+++ chk/modules/testcase/support.py
@@ -1,11 +1,11 @@
 """
-test_spec related support services
+testcase related support services
 """
 from cerberus import validator
 from chk.infrastructure.exception import err_message, messages
-from chk.modules.test_spec.constants import TestSpecConfigNode, ExecuteConfigNode
+from chk.modules.testcase.constants import TestSpecConfigNode, ExecuteConfigNode
 from chk.modules.http.constants import RequestConfigNode
-from chk.modules.test_spec.validation_rules import test_spec_schema
+from chk.modules.testcase.validation_rules import testcase_schema
 
 
 class TestSpecMixin:
@@ -13,19 +13,19 @@ class TestSpecMixin:
     Mixin for version spec. for v0.7.2
     """
 
-    def __init_test_spec_mixin(self):
+    def __init_testcase_mixin(self):
         self.in_file = True
         
     def is_request_infile(self) -> bool:
         return self.in_file;
 
-    def test_spec_validated(self) -> dict[str, str]:
+    def testcase_validated(self) -> dict[str, str]:
         """
         Validate the schema against config
         """
         try:
-            test_spec_doc = self.test_spec_as_dict()
-            if not self.validator.validate(test_spec_doc, test_spec_schema):
+            testcase_doc = self.testcase_as_dict()
+            if not self.validator.validate(testcase_doc, testcase_schema):
                 raise SystemExit(err_message('fatal.V0006', extra=self.validator.errors))
 
             # validate request if it exists in-file
@@ -34,9 +34,9 @@ class TestSpecMixin:
         except validator.DocumentError as doc_err:
             raise SystemExit(err_message('fatal.V0001', extra=doc_err)) from doc_err
 
-        return test_spec_doc  # or is a success
+        return testcase_doc  # or is a success
 
-    def test_spec_as_dict(self) -> dict[str, str]:
+    def testcase_as_dict(self) -> dict[str, str]:
         """
         Get version string
         """
@@ -62,7 +62,7 @@ class TestSpecMixin:
         if in_file_request is out_file_request:
             raise SystemExit(err_message('fatal.V0020', extra={'spec': {'execute': {'file': '...'}}}))
 
-        self.__init_test_spec_mixin()
+        self.__init_testcase_mixin()
 
         if in_file_request:
             self.request_validated()
