+++ tests/resources/storage/sample_config/pass_cases/UserCreate/01_UserCreateSpec.chk
@@ -1,4 +1,4 @@
-version: 'default:test-spec:0.7.2'
+version: 'default:testcase:0.7.2'
 
 variables:
   $Code: ~
