+++ tests/resources/storage/sample_config/pass_cases/UserCreate/02_UserCreateSpecWithRequest.chk
@@ -1,4 +1,4 @@
-version: 'default:test-spec:0.7.2'
+version: 'default:testcase:0.7.2'
 
 variables:
   Name: 'Morpheus'
